chore: rename codex-command to codex-shell-command
This addresses some post-merge feedback on https://github.com/openai/codex/pull/11361:

- crate rename
- reuse `detect_shell_type()` utility
