+++ codex-rs/core/src/lib.rs
@@ -56,6 +56,7 @@ mod proposed_plan_parser;
 mod sandbox_tags;
 pub mod sandboxing;
 mod session_prefix;
+mod shell_detect;
 mod stream_events_utils;
 mod tagged_block_parser;
 mod text_encoding;
@@ -129,11 +130,11 @@ mod state;
 mod tasks;
 mod user_shell_command;
 pub mod util;
-pub use codex_command::bash;
-pub use codex_command::is_dangerous_command;
-pub use codex_command::is_safe_command;
-pub use codex_command::parse_command;
-pub use codex_command::powershell;
+pub use codex_shell_command::bash;
+pub use codex_shell_command::is_dangerous_command;
+pub use codex_shell_command::is_safe_command;
+pub use codex_shell_command::parse_command;
+pub use codex_shell_command::powershell;
 
 pub use apply_patch::CODEX_APPLY_PATCH_ARG1;
 pub use client::X_CODEX_TURN_METADATA_HEADER;
