+++ codex-rs/shell-command/src/bash.rs
