+++ codex-rs/shell-command/src/command_safety/is_safe_command.rs
