+++ codex-rs/shell-command/src/command_safety/mod.rs
