+++ codex-rs/shell-command/src/command_safety/powershell_parser.ps1
