+++ codex-rs/shell-command/src/powershell.rs
