+++ codex-rs/shell-command/src/shell_detect.rs
